requirements.txt
streamlit
gdown
numpy
Pillow
tensorflow